+++ src/components/auth/Register.jsx
